refactor: introduce new activity log schemas for system audit trail, risk alert log, integration audit log, user login history, cron job run, notification log, and network access log
