+++ packages/data/schema/system/src/model/activity-logs/activity-logs.schema.ts
@@ -1,0 +1,2 @@
+// Re-export all schemas from individual files for backward compatibility
+export * from './index';
